Fully Working Camera for Game Activity
Finally got it to work. I feel dead but victorious.

Users will be able to tap each image of the game or the custom pictures and be able to take a photo as they wish. This should also be able to use any photo saved locally to the device if the correct file path is saved.